Run array specialization in a sperate pass.
